requirements.txt
asgiref==3.8.1
attrs==24.2.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
cors==1.0.1
Django==4.2.15
django-request==1.6.3
filelock==3.15.4
flask==3.0.3
Flask-Cors==4.0.1
flatbuffers==24.3.25
future==1.0.0
gevent==24.2.1
greenlet==3.0.3
humanfriendly==10.0
idna==3.7
imageio==2.35.1
importlib-metadata==8.4.0
itsdangerous==2.2.0
jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lazy-loader==0.4
llvmlite==0.43.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.1
onnxruntime==1.19.0
opencv-python-headless==4.10.0.84
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
protobuf==5.27.3
PyMatting==1.1.12
PySocks==1.7.1
python-dateutil==2.9.0.post0
referencing==0.35.1
requests==2.32.3
requests-file==2.1.0
rpds-py==0.20.0
scikit-image==0.24.0
scipy==1.13.1
six==1.16.0
sqlparse==0.5.1
sympy==1.13.2
tifffile==2024.8.10
tldextract==5.1.2
tqdm==4.66.5
typing-extensions==4.12.2
urllib3==2.2.2
werkzeug==3.0.4
zipp==3.20.0
zope.event==5.0
zope.interface==7.0.1
